requirements.txt
pytest>=3.6.1,<3.7.0
web3>=4.4.0,<4.5.0
pyyaml>=3.12,<3.13
requests>=2.19,<2.20
python-dateutil>=2.7,<2.8
